Делит проект на модули
